Changed some things a bit
Changed where data files are located (Now using AppData/Roaming folder, and not something relative to the executable).

Also changed how anime moving works. If files are on the same drive, will just move them, since it's faster, but if not on the same drive, will copy the source to the destination, then delete the source, since it's the same speed as moving and is a bit more fail-proof.